Add files via upload
The files beginning with Eman-2 are all part of the frontal walking animation. The second number indicates the frame number (1 is excluded because that's for him standing still); you can ignore the third number, as it's just the version number. Eman-3 is the left walking animation, Eman-4 is the right walking animation, and Eman-5 is the rear walking animation.